Refactor the test for load_study_cohort
By patching each Pandas read function with a mock object, we can side
step the need to create fixtures (in this case, the files in
the tests/test_data directory). We can also quickly test two unhappy
paths: when the user passes a path to an unsupported file type; and when
the user passes a path to a supported file type that isn't supported
correctly.