Ensure the input CSV has 20 rows
Previously, the input CSV had eight rows: that is, the Cartesian product
of three columns, each with two categories. So few rows will almost
always be redacted, regardless of the redaction strategy. By increasing
the input CSV to 20 rows, and by ensuring that each category has 10 rows
apiece, we ensure the rows are not redacted.