Replace age_band (categorical) with bmi (float)
We don't really need two categorical variables in the input CSV; that
is, sex and age_band. However, we do need a float variable, such as bmi,
so let's replace age_band with bmi.

Rather than simulate "true" BMI values, we simply provide a range. We
also increase the number of rows in the input CSV from 20 to 100. We do
this to ensure that rows are not redacted.