Supply more complex test data
Following @inglesp's comment on #51, this supplies more complex test
data to `_series_with_interval_index_to_histogram`. In doing so, we
guard against an implementation that, for example, reduces the bin edges
to a single value from the left interval (rather than all values from
the left interval).

There are other tests that would benefit from more complex test data.
However, I feel that it's a better use of time to improve them
as-and-when required. For the same reason, I'll close #51.

Closes #51