Pass file name to report
The `graph` property is used to create an image tag in the report. When
this property is a relative file path, it is interpreted as being
relative to the report, which is incorrect. We change it to a file name,
so the image is correctly located within the same directory as the
report.

Fixes #54