Test `main`
Rather than fixing `convert_config`, we're going to replace it. This is
because configuration shouldn't be stored in a JSON file; it should be
stored in a study's *project.yaml* instead.

`convert_config` is called from `main`, so we're going to test the
configuration-handling aspects of this function before replacing
`convert_config` in the next commit.