Add a test that recreates the error
The error occurs when a path to a csv.gz file is passed to
`make_report`. This function checks this path's suffix and, if equal to
".csv.gz", raises an error. However, the suffix is the *final* component
of the path, which is ".gz" in the case of a path to a ".csv.gz" file.